BA-1809 Implement remaining gitlab suggestions
