--- packages/authentication/modules/profile/useCurrentProfile/__tests__/__utils__/TestComponent/index.tsx
@@ -1,17 +1,0 @@
-import useCurrentProfile from '../../..'
-
-const TestComponent = () => {
-  const profile = useCurrentProfile().currentProfile
-
-  if (!profile) {
-    return null
-  }
-
-  return (
-    <>
-      <span id="profile-id">{profile.id}</span>
-    </>
-  )
-}
-
-export default TestComponent
